Bindings: Moved bindings core and basic shapes to the separate file, imported bindings in annotations.
