+++ js/masters/modules/annotations-advanced.src.js
@@ -16,3 +16,4 @@ import '../../annotations/types/Fibonacci.js';
 import '../../annotations/types/Pitchfork.js';
 import '../../annotations/types/VerticalLine.js';
 import '../../annotations/types/Measure.js';
+import '../../annotations/navigationBindings.js';
