+++ js/masters/modules/annotations.src.js
@@ -8,3 +8,4 @@
  */
 'use strict';
 import '../../annotations/annotations.src.js';
+import '../../annotations/navigationBindings.js';
